+++ src/main/java/refinedstorage/tile/CraftingTaskScheduler.java
@@ -8,32 +8,29 @@ import refinedstorage.tile.autocrafting.CraftingPattern;
 public class CraftingTaskScheduler {
     public static String NBT_SCHEDULED = "CraftingTaskScheduled";
 
-    private ItemStack scheduledFor;
+    private ItemStack scheduledItem;
 
     public boolean canSchedule(int compare, ItemStack item) {
-        // We can only reschedule if:
-        // - we didn't schedule anything before
-        // - the item we can't to schedule is another item
-        return scheduledFor == null || !RefinedStorageUtils.compareStack(scheduledFor, item, compare);
+        return scheduledItem == null || !RefinedStorageUtils.compareStack(scheduledItem, item, compare);
     }
 
     public void schedule(TileController controller, int compare, ItemStack item) {
         CraftingPattern pattern = controller.getPattern(item, compare);
 
         if (pattern != null) {
-            scheduledFor = item;
+            scheduledItem = item;
 
             controller.addCraftingTask(pattern);
         }
     }
 
     public void resetSchedule() {
-        this.scheduledFor = null;
+        this.scheduledItem = null;
     }
 
     public void writeToNBT(NBTTagCompound tag) {
-        if (scheduledFor != null) {
-            tag.setTag(NBT_SCHEDULED, scheduledFor.serializeNBT());
+        if (scheduledItem != null) {
+            tag.setTag(NBT_SCHEDULED, scheduledItem.serializeNBT());
         } else {
             tag.removeTag(NBT_SCHEDULED);
         }
@@ -41,7 +38,7 @@ public class CraftingTaskScheduler {
 
     public void read(NBTTagCompound tag) {
         if (tag.hasKey(NBT_SCHEDULED)) {
-            scheduledFor = ItemStack.loadItemStackFromNBT(tag.getCompoundTag(NBT_SCHEDULED));
+            scheduledItem = ItemStack.loadItemStackFromNBT(tag.getCompoundTag(NBT_SCHEDULED));
         }
     }
 }